+++ spec/text_spec.lua
@@ -165,34 +165,6 @@ three
 
   describe("fill()/wrap()", function()
 
-    it("word-wraps a text", function()
-      assert.equal([[
-It is often said of Lua
-that it does not include
-batteries. That is because
-the goal of Lua is to
-produce a lean expressive
-language that will be
-used on all sorts of
-machines, (some of which
-don't even have hierarchical
-filesystems). The Lua
-language is the equivalent
-of an operating system
-kernel; the creators
-of Lua do not see it
-as their responsibility
-to create a full software
-ecosystem around the
-language. That is the
-role of the community.
-]], text.fill("It is often said of Lua that it does not include batteries. That is because the goal of Lua is to produce a lean expressive language that will be used on all sorts of machines, (some of which don't even have hierarchical filesystems). The Lua language is the equivalent of an operating system kernel; the creators of Lua do not see it as their responsibility to create a full software ecosystem around the language. That is the role of the community.", 20))
-    end)
-
-    it("wraps single letters", function()
-      assert.same({"a"}, text.wrap("a"))
-    end)
-
     it("wraps width over limit", function()
       assert.same({
         "abc",
@@ -205,15 +177,72 @@ role of the community.
         "abc",
         "def"
       }, text.wrap("abc def", 3))
+      assert.same({
+        "a c",
+        "d f"
+      }, text.wrap("a c d f", 3))
+    end)
+
+    it("wraps single letters", function()
+      assert.same({"a"}, text.wrap("a"))
+    end)
+
+    it("wraps empty strings", function()
+      assert.same({""}, text.wrap(""))
+      assert.same({""}, text.wrap("    "))
+    end)
+
+    it("handles leading/trailing whitespace", function()
+      assert.same({"hello"}, text.wrap("     hello     ", 10))
+      assert.same({"hello"}, text.wrap("     hello     ", 2))
+      assert.same({"he", "ll", "o"}, text.wrap("     hello     ", 2, true))
+    end)
+
+    it("handles line-breaks", function()
+      assert.same({"Hello", "Dolly"}, text.wrap("Hello\nDolly", 10))
+      assert.same({"Hello Dolly"}, text.wrap("Hello\nDolly", 20))
     end)
 
     it("doesn't split on accented characters", function()
       assert.same({"àbcdéfghîj"}, text.wrap("àbcdéfghîj"))
     end)
 
+    it("word-wraps a text", function()
+      local binstring = require("luassert.formatters.binarystring")
+      assert:add_formatter(binstring)
+      assert.equal([[
+It is often said of
+Lua that it does not
+include batteries.
+That is because the
+goal of Lua is to
+produce a lean
+expressive language
+that will be used on
+all sorts of
+machines, (some of
+which don't even
+have hierarchical
+filesystems). The
+Lua language is the
+equivalent of an
+operating system
+kernel; the creators
+of Lua do not see it
+as their
+responsibility to
+create a full
+software ecosystem
+around the language.
+That is the role of
+the community.
+]], text.fill("It is often said of Lua that it does not include batteries. That is because the goal of Lua is to produce a lean expressive language that will be used on all sorts of machines, (some of which don't even have hierarchical filesystems). The Lua language is the equivalent of an operating system kernel; the creators of Lua do not see it as their responsibility to create a full software ecosystem around the language. That is the role of the community.", 20))
+    end)
+
+
     it("generic wrap test", function()
       local t = [[
-hello "world" 'this' -is- a bb ccc dddd test... but wouldn't it pass??? final. word-that-can-be-broken
+hello "world" 'this' -is- a bb      ccc dddd test... but wouldn't it pass??? final. word-that-can-be-broken
 ]]
 
       assert.same({
@@ -221,16 +250,61 @@ hello "world" 'this' -is- a bb ccc dddd test... but wouldn't it pass??? final. w
         '"world"',
         "'this'",
         "-is-",
-        "a bb",
+        "a",
+        "bb",
         "ccc",
         "dddd",
         "test...",
         "but",
         "wouldn't",
-        "it pass???",
+        "it",
+        "pass???",
         "final.",
         "word-that-can-be-broken",
       }, text.wrap(t, 3))
+    end)
+
+    it("generic wrap test, with overflow breaking", function()
+      local t = [[
+hello "world" 'this' -is- a bb      ccc dddd test... but wouldn't it pass??? final. word-that-can-be-broken
+]]
+
+      assert.same({
+        "hel",
+        "lo",
+        '"wo',
+        'rld',
+        '"',
+        "'th",
+        "is'",
+        "-is",
+        "- a",
+        "bb",
+        "ccc",
+        "ddd",
+        "d",
+        "tes",
+        "t..",
+        ".",
+        "but",
+        "wou",
+        "ldn",
+        "'t",
+        "it",
+        "pas",
+        "s??",
+        "?",
+        "fin",
+        "al.",
+        "wor",
+        "d-t",
+        "hat",
+        "-ca",
+        "n-b",
+        "e-b",
+        "rok",
+        "en",
+      }, text.wrap(t, 3, true))
     end)
 
   end)