+++ CHANGELOG.md
@@ -13,6 +13,7 @@ see [CONTRIBUTING.md](CONTRIBUTING.md#release-instructions-for-a-new-version) fo
    [#396](https://github.com/lunarmodules/Penlight/pull/396)
  - fix: `text.dedent` didn't handle declining indents nor empty lines
    [#402](https://github.com/lunarmodules/Penlight/pull/402)
+<<<<<<< HEAD
  - fix: `dir.getfiles`, `dir.getdirectories`, and `dir.getallfiles` now have the
    directory optional, as was already documented
    [#405](https://github.com/lunarmodules/Penlight/pull/405)
@@ -20,6 +21,14 @@ see [CONTRIBUTING.md](CONTRIBUTING.md#release-instructions-for-a-new-version) fo
    also other functions now take a range. [#404](https://github.com/lunarmodules/Penlight/pull/404)
  - fix: `lapp` enums allow [patterns magic characters](https://www.lua.org/pil/20.2.html)
    [#393](https://github.com/lunarmodules/Penlight/pull/393)
+=======
+ - fix: `text.wrap` and `text.fill` numerous fixes for handling whitespace,
+   accented characters, honouring width, etc.
+   [#400](https://github.com/lunarmodules/Penlight/pull/400)
+ - feat: `text.wrap` and `text.fill` have a new parameter to forcefully break words
+   longer than the width given.
+   [#400](https://github.com/lunarmodules/Penlight/pull/400)
+>>>>>>> 7b3efef4 (feat(text) wrap/fill can break long words, honours width)
 
 ## 1.11.0 (2021-08-18)